pyproject.toml
[project]
name = "bonsaizmqexamples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyzmq>=27.1.0",
]
